refactor: 예약 추가 시 시간값을 파싱하도록 수정
