+++ src/main/java/roomescape/dto/reservation/ReservationRequestDTO.java
@@ -1,11 +1,12 @@
-package roomescape.dto;
+package roomescape.dto.reservation;
 
 import jakarta.validation.constraints.NotEmpty;
 import lombok.AccessLevel;
 import lombok.NoArgsConstructor;
 
 @NoArgsConstructor(access = AccessLevel.PRIVATE)
+@NotEmpty
 public class ReservationRequestDTO {
-  	public record AddReservationRequest(@NotEmpty String name, @NotEmpty String date, @NotEmpty String time) {}
+	public record AddReservationRequest(String name, String date, Long time_id) {
 	}
 }
